- 2.6.11 update -- the main thing

+++ linux-libc-headers/trunk/include/linux/mtd/doc2000.h
@@ -88,8 +88,8 @@
 #define WriteDOC_(d, adr, reg)  do{ *(volatile __u16 *)(((unsigned long)adr)+((reg)<<1)) = (__u16)d; wmb();} while(0)
 #define DOC_IOREMAP_LEN 0x4000
 #else
-#define ReadDOC_(adr, reg)      readb((void *)(((unsigned long)adr) + (reg)))
-#define WriteDOC_(d, adr, reg)  writeb(d, (void *)(((unsigned long)adr) + (reg)))
+#define ReadDOC_(adr, reg)      readb((void *)(adr) + (reg))
+#define WriteDOC_(d, adr, reg)  writeb(d, (void *)(adr) + (reg))
 #define DOC_IOREMAP_LEN 0x2000
 
 #endif
